+++ frontend/features/mainNavigation/blocks/index.ts
